mvc model
